organise code placing libraries directly inside of index.js, rewrite code for displayCaategories.js to use flickity methods
